+++ include/dislua/const.hpp
@@ -3,7 +3,7 @@
 
 // MIT License
 
-// Copyright (c) 2020-2021 Vitaliy Vorobets
+// Copyright (c) 2020-2022 Vitaliy Vorobets
 
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -34,7 +34,7 @@
  * @brief disluapp version.
  * Format: MAJOR.MINOR.PATCH
  */
-#define DISLUA_VERSION 111L
+#define DISLUA_VERSION 112L
 
 namespace dislua {
 using uchar = unsigned char;