+++ include/dislua/lj/ljparser.hpp
@@ -3,7 +3,7 @@
 
 // MIT License
 
-// Copyright (c) 2020-2021 Vitaliy Vorobets
+// Copyright (c) 2020-2022 Vitaliy Vorobets
 
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -58,7 +58,7 @@ class parser : public dump_info {
   std::string read_to_zero() {
     std::string s;
     char c;
-    while (c = buf->read<char>(), c != '\0')
+    while (c = buf.read<char>(), c != '\0')
       s.push_back(c);
     return s;
   }
@@ -67,7 +67,7 @@ class parser : public dump_info {
   T read_with_uleb128() {
     constexpr size_t num = sizeof(T) / sizeof(uleb128);
     uleb128 vals[num];
-    buf->read_uleb128(vals, num);
+    buf.read_uleb128(vals, num);
     return std::bit_cast<T>(vals);
   }
 
@@ -81,16 +81,16 @@ class parser : public dump_info {
 
   /// Read
   void read_header() {
-    if (buf->read() != header::HEAD1
-     || buf->read() != header::HEAD2
-     || buf->read() != header::HEAD3)
+    if (buf.read() != header::HEAD1
+     || buf.read() != header::HEAD2
+     || buf.read() != header::HEAD3)
       throw std::runtime_error("LuaJIT: Invalid header.");
 
-    version = static_cast<uint>(buf->read());
+    version = static_cast<uint>(buf.read());
     if (version != 2 && version != 1)
       throw std::runtime_error("LuaJIT: Unknown version.");
 
-    uleb128 flags = buf->read_uleb128();
+    uleb128 flags = buf.read_uleb128();
     header.flags = flags;
 
     flags &= ~dump_flags::be;
@@ -102,28 +102,28 @@ class parser : public dump_info {
       throw std::runtime_error("LuaJIT: Unknown header flags.");
 
     if ((header.flags & dump_flags::strip) == 0) {
-      uleb128 len = buf->read_uleb128();
+      uleb128 len = buf.read_uleb128();
       header.debug_name.resize(len);
-      buf->read(header.debug_name.begin(), header.debug_name.end());
+      buf.read(header.debug_name.begin(), header.debug_name.end());
     }
   }
 
   void read_bc_instructions(proto &pt, uleb128 size) {
     while (size--) {
-      instruction ins = buf->read<instruction>();
+      instruction ins = buf.read<instruction>();
       pt.ins.push_back(ins);
     }
   }
 
   void read_uv(proto &pt, uchar size) {
     while (size--) {
-      ushort uv_data = buf->read<ushort>();
+      ushort uv_data = buf.read<ushort>();
       pt.uv.push_back(uv_data);
     }
   }
 
   table_val_t read_ktabk() {
-    uleb128 type = buf->read_uleb128();
+    uleb128 type = buf.read_uleb128();
     table_val_t value;
 
     switch (type) {
@@ -137,20 +137,20 @@ class parser : public dump_info {
       value = true;
       break;
     case ktab::integer: {
-      uleb128 v = buf->read_uleb128();
+      uleb128 v = buf.read_uleb128();
       value = static_cast<leb128>(v);
       break;
     }
     case ktab::number: {
       uleb128 vals[2] = {0};
-      buf->read_uleb128(vals, 2);
+      buf.read_uleb128(vals, 2);
       value = std::bit_cast<double>(vals);
       break;
     }
     default: {
       std::string v;
       v.resize(type - ktab::string);
-      buf->read(v.begin(), v.end());
+      buf.read(v.begin(), v.end());
       value = v;
       break;
     }
@@ -161,8 +161,8 @@ class parser : public dump_info {
 
   table_t read_ktab() {
     table_t table;
-    uleb128 narray = buf->read_uleb128();
-    uleb128 nhash = buf->read_uleb128();
+    uleb128 narray = buf.read_uleb128();
+    uleb128 nhash = buf.read_uleb128();
 
     for (uleb128 i = 0; i < narray; ++i) {
       table_val_t value = read_ktabk();
@@ -180,7 +180,7 @@ class parser : public dump_info {
   void read_kgc(proto &pt, uleb128 size) {
     uleb128 type;
     while (size--) {
-      type = buf->read_uleb128();
+      type = buf.read_uleb128();
 
       switch (type) {
       case kgc::child:
@@ -207,7 +207,7 @@ class parser : public dump_info {
       default: {
         uleb128 length = type - kgc::string;
         std::string str(length, '\0');
-        buf->read(str.begin(), str.end());
+        buf.read(str.begin(), str.end());
         pt.kgc.emplace_back(str);
         break;
       }
@@ -217,12 +217,12 @@ class parser : public dump_info {
 
   void read_knum(proto &pt, uleb128 size) {
     while (size--) {
-      bool isnum = buf->read(false) & 1;
+      bool isnum = buf.read(false) & 1;
 
-      uleb128 result[2] = {buf->read_uleb128_33(), 0};
+      uleb128 result[2] = {buf.read_uleb128_33(), 0};
       double v;
       if (isnum) {
-        result[1] = buf->read_uleb128();
+        result[1] = buf.read_uleb128();
         v = std::bit_cast<double>(result);
       } else
         v = static_cast<double>(static_cast<leb128>(result[0]));
@@ -234,11 +234,11 @@ class parser : public dump_info {
     while (size--) {
       uint line;
       if (pt.numline >= 1 << 16)
-        line = buf->read<uint>();
+        line = buf.read<uint>();
       else if (pt.numline >= 1 << 8)
-        line = static_cast<uint>(buf->read<ushort>());
+        line = static_cast<uint>(buf.read<ushort>());
       else
-        line = static_cast<uint>(buf->read());
+        line = static_cast<uint>(buf.read());
       pt.lineinfo.push_back(pt.firstline + line);
     }
   }
@@ -251,17 +251,17 @@ class parser : public dump_info {
   void read_varname(proto &pt) {
     size_t last = 0;
     uchar type;
-    while (type = buf->read(), type != varnames::end) {
+    while (type = buf.read(), type != varnames::end) {
       varname info = {};
       info.type = type;
 
       if (type >= varnames::MAX) {
-        buf->iread--;
+        buf.iread--;
         info.name = read_to_zero();
       }
 
-      last = info.start = last + static_cast<size_t>(buf->read_uleb128());
-      info.end = info.start + static_cast<size_t>(buf->read_uleb128());
+      last = info.start = last + static_cast<size_t>(buf.read_uleb128());
+      info.end = info.start + static_cast<size_t>(buf.read_uleb128());
 
       pt.varnames.push_back(info);
     }
@@ -271,14 +271,14 @@ class parser : public dump_info {
     uleb128 sizekgc, sizekn, sizebc;
     uchar sizeuv;
 
-    pt.flags = buf->read();
-    pt.numparams = buf->read();
-    pt.framesize = buf->read();
-    sizeuv = buf->read();
+    pt.flags = buf.read();
+    pt.numparams = buf.read();
+    pt.framesize = buf.read();
+    sizeuv = buf.read();
 
-    sizekgc = buf->read_uleb128();
-    sizekn = buf->read_uleb128();
-    sizebc = buf->read_uleb128();
+    sizekgc = buf.read_uleb128();
+    sizekn = buf.read_uleb128();
+    sizebc = buf.read_uleb128();
 
     uchar flags = pt.flags;
     flags &= ~proto_flags::child;
@@ -291,10 +291,10 @@ class parser : public dump_info {
 
     uleb128 sizedbg = 0;
     if ((header.flags & dump_flags::strip) == 0) {
-      sizedbg = buf->read_uleb128();
+      sizedbg = buf.read_uleb128();
       if (sizedbg) {
-        pt.firstline = buf->read_uleb128();
-        pt.numline = buf->read_uleb128();
+        pt.firstline = buf.read_uleb128();
+        pt.numline = buf.read_uleb128();
       }
     }
 
@@ -303,28 +303,28 @@ class parser : public dump_info {
     read_kgc(pt, sizekgc);
     read_knum(pt, sizekn);
 
-    size_t sizedbg_read = buf->iread;
+    size_t sizedbg_read = buf.iread;
     if (sizedbg) {
       read_lineinfo(pt, sizebc);
       read_uvname(pt, sizeuv);
       read_varname(pt);
     }
-    if (buf->iread - sizedbg_read != sizedbg)
+    if (buf.iread - sizedbg_read != sizedbg)
       throw std::runtime_error("LuaJIT: Invalid debuginfo size.");
   }
 
   /// Write
   void write_header() {
-    buf->write(header::HEAD1);
-    buf->write(header::HEAD2);
-    buf->write(header::HEAD3);
+    buf.write(header::HEAD1);
+    buf.write(header::HEAD2);
+    buf.write(header::HEAD3);
 
-    buf->write(static_cast<uchar>(version));
-    buf->write_uleb128(header.flags);
+    buf.write(static_cast<uchar>(version));
+    buf.write_uleb128(header.flags);
 
     if ((header.flags & dump_flags::strip) == 0) {
-      buf->write_uleb128(static_cast<uleb128>(header.debug_name.size()));
-      buf->write(header.debug_name.begin(), header.debug_name.end());
+      buf.write_uleb128(static_cast<uleb128>(header.debug_name.size()));
+      buf.write(header.debug_name.begin(), header.debug_name.end());
     }
   }
 
@@ -384,7 +384,7 @@ class parser : public dump_info {
 
   void write_kgc(proto &pt, buffer &ptbuf) {
     for (kgc_t &kgc: pt.kgc)
-      std::visit(detail::overloaded{[&](const proto &) {
+      std::visit(detail::overloaded{[&](const proto_id &) {
                                       ptbuf.write_uleb128(kgc::child);
                                     },
                                     [&](table_t arg) {
@@ -412,7 +412,8 @@ class parser : public dump_info {
 
   void write_knum(proto &pt, buffer &ptbuf) {
     for (double &val: pt.knum) {
-      if (detail::almost_equal(std::fmod(val, 1), 0.0, 2))
+      double sval = static_cast<double>(static_cast<int>(val)); // signed value
+      if (detail::almost_equal(val, sval, 2))
         ptbuf.write_uleb128_33(static_cast<uleb128>(val));
       else {
         uleb128 v[2] = {0};
@@ -495,55 +496,55 @@ class parser : public dump_info {
       ptbuf.write(ptdebug);
 
     size_t len = ptbuf.iwrite;
-    buf->write_uleb128(static_cast<uleb128>(len));
-    buf->write(ptbuf);
+    buf.write_uleb128(static_cast<uleb128>(len));
+    buf.write(ptbuf);
   }
 
-  std::vector<proto> temp_protos;
+  std::vector<proto_id> temp_protos;
 
 public:
   explicit parser(const buffer &_buf = {}) : dump_info(_buf) {}
   explicit parser(const dump_info &rv) : dump_info(rv) {}
 
   void read() override {
-    buf->reset_indices();
+    buf.reset_indices();
 
     read_header();
-    while (buf->iread < buf->size()) {
-      if (!buf->read(false))
+    while (buf.iread < buf.size()) {
+      if (!buf.read(false))
         break;
 
-      uleb128 len = buf->read_uleb128();
+      uleb128 len = buf.read_uleb128();
       if (!len)
         throw std::runtime_error("LuaJIT: Prototype size == 0.");
 
       proto pt;
-      size_t start = buf->iread;
+      size_t start = buf.iread;
       read_proto(pt);
-      if (buf->iread - start != len)
+      if (buf.iread - start != len)
         throw std::runtime_error("LuaJIT: Different prototype sizes.");
-      temp_protos.push_back(pt);
+      temp_protos.emplace_back(temp_protos.size());
       protos.push_back(pt);
     }
 
     if (temp_protos.size() != 1)
-      throw std::runtime_error("LuaJit: Invalid prototype stack.");
-    if (buf->iread + 1 > buf->size())
+      throw std::runtime_error("LuaJIT: Invalid prototype stack.");
+    if (buf.iread + 1 > buf.size())
       throw std::runtime_error("LuaJIT: Something not read.");
     temp_protos.clear();
 
-    buf->reset_indices();
+    buf.reset_indices();
   }
 
   void write() override {
-    buf->reset();
+    buf.reset();
 
     write_header();
     for (proto &pt: protos)
       write_proto(pt);
-    buf->write('\0');
+    buf.write('\0');
 
-    buf->reset_indices();
+    buf.reset_indices();
   }
 
   compilers compiler() override {